fix(web): resolve nextjs build errors related to Html tag and Suspense
ビルドエラーを修正:
- カスタム404ページ (not-found.tsx) を追加して Html タグエラーを解消
- useSearchParams を使用するページ (/auth/signin, /auth/error) で Suspense を使用するようにリファクタリング
